% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralphabet.R
\name{ralphabet}
\alias{ralphabet}
\title{Generates a random letter}
\usage{
ralphabet()
}
\value{
A random letter form the alphabet.
}
\description{
Generates a random letter
}
\examples{
set.seed(1)
ralphabet()
"Y"
}
